[webform] Make $element['webform_component'] the authoritative storage for properties.
The new values for the properties are now always kept inside the
$element['#webform_component'] sub-array. The rest of the $element is
only used to store the renderable element and form_builder's own
metadata. Thus we cease to make assumptions about how a webform
component is rendered.

This commit also removes a lot of workarounds related to the previous
behavior.

This commit is likely to fix at least #2463259 and #2144253.